+++ discopy/balanced.py
@@ -340,6 +340,7 @@ class DualRail(Functor):
 
 CMap = cmap.CMap[Diagram]
 Hypergraph = hypergraph.Hypergraph[Diagram]
+ToHypergraph = Diagram.ToHypergraph
 Diagram.DualRail = DualRail
 Layer = Diagram.Layer
 Id = Diagram.id