+++ discopy/grammar/pregroup.py
@@ -321,5 +321,7 @@ TermBase, Constant, Variable, Application, Abstraction = (
     Diagram.Application, Diagram.Abstraction)
 Layer = Diagram.Layer
 Id = Diagram.id
+ToHypergraph = Diagram.ToHypergraph
+ToMap = Diagram.ToMap