+++ discopy/ribbon.py
@@ -345,6 +345,7 @@ TermBase, Constant, Variable, Application, Abstraction = (
     Diagram.Application, Diagram.Abstraction)
 Layer = Diagram.Layer
 Id = Diagram.id
+ToHypergraph = Diagram.ToHypergraph
 
 
 class Equation(pivotal.Equation):